+++ src/Api/MergeRequests.php
@@ -61,6 +61,8 @@ class MergeRequests extends AbstractApi
      *     @var string             $labels         return merge requests matching a comma separated list of labels
      *     @var \DateTimeInterface $created_after  return merge requests created after the given time (inclusive)
      *     @var \DateTimeInterface $created_before return merge requests created before the given time (inclusive)
+     *     @var int                $reviewer_id    return merge requests which have the user as a reviewer with the given user id
+     *     @var bool               $wip            return only draft merge requests (true) or only non-draft merge requests (false)
      * }
      *
      * @throws UndefinedOptionsException if an option name is undefined
@@ -136,6 +138,13 @@ class MergeRequests extends AbstractApi
                 return \count($value) === \count(\array_filter($value, 'is_int'));
             })
         ;
+        $resolver->setDefined('reviewer_id')
+            ->setAllowedTypes('reviewer_id', 'integer');
+        $resolver->setDefined('wip')
+            ->setAllowedTypes('wip', 'boolean')
+            ->addNormalizer('wip', static function ($resolver, $wip) {
+                return $wip ? 'yes' : 'no';
+            });
 
         $path = null === $project_id ? 'merge_requests' : $this->getProjectPath($project_id, 'merge_requests');